+++ Assets/LDtkUnity/Runtime/Tools/EllipseTest.cs
@@ -5,9 +5,14 @@ namespace LDtkUnity
 {
     public class EllipseTest : MonoBehaviour
     {
+        [SerializeField, Range(0,10)] private float thickness;
+        [SerializeField, Range(0,1)] private float alpha;
+        [SerializeField, Range(0,100)] private int pointCount;
+        
         private void OnDrawGizmos()
         {
-            GizmoUtil.DrawEllipse(transform.position, transform.localScale);
+            GizmoUtil.DrawAAEllipse(transform.position, transform.localScale, pointCount, thickness, alpha);
+            //GizmoUtil.DrawEllipse(transform.position, transform.localScale);
         }
     }
 }